pkg/ebpf,pkg/process: fix linting issues
